e/manage_car: comment out all API calls
This "example code" definitely demonstrates how to use the library, but
is surprisingly wreckless in its functionality. It is not the sort of
example you should blindly run, which is almost certainly what some
users are likely to do.

Rather than try to pick and choose which of the possibly-destructive
actions to comment out (such as auto-park), just don't do anything but
print out whether or not the API (mobile app access) is enabled and
exit. The existing API calls are all still here for developer viewing,
but commented out.